fix: cache instance cleanup
